dinner-picker-backend/backendV1/requirements/dev.txt
certifi==2018.4.16
chardet==3.0.4
click==6.7
coverage==4.5.1
Django==2.0.7
djangorestframework==3.8.2
Flask==1.0.2
idna==2.7
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
psycopg2==2.7.5
psycopg2-binary==2.7.5
Pygments==2.2.0
pytz==2018.5
requests==2.19.1
urllib3==1.23
Werkzeug==0.14.1
